+++ resources/stubs/module/jobs.php
@@ -1,2 +1,2 @@
-<?php 
-['seed'            => '{{vendor|studly_case}}\{{module_slug|studly_case}}Module\Listener\Command\Seed'];
+<?php
+['seed'            => '{{vendor|studly_case}}\{{module_name}}Module\Listener\Command\Seed'];