+++ resources/stubs/module/routes.php
@@ -1,2 +1,2 @@
-<?php 
-['admin/{{module_slug|lower}}'            => '{{vendor|studly_case}}\{{module_slug|studly_case}}Module\Http\Controller\Admin\{{module_slug|studly_case}}Controller@index'];
+<?php
+['admin/{{module_name|lower}}'            => '{{vendor|studly_case}}\{{module_name}}Module\Http\Controller\Admin\{{module_name}}Controller@index'];