+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+1.1.1
+  date: 2016-10-12
+  changes:
+  - Fix landing page routes
+
 1.1.0
   date: 2016-10-03
   changes: